refactor: update import paths for BaseService and enhance user service query methods for improved readability and consistency
