+++ src/services/v2/account/account-settings-service.js
@@ -1,4 +1,4 @@
-import { BaseService } from '../base'
+import { BaseService } from '../base/BaseService'
 
 export class AccountSettingsService extends BaseService {
   constructor() {
