+++ src/services/v2/base/cache/CacheConfig.js
@@ -8,13 +8,13 @@ export const CACHE_TYPES = {
 export const cacheConfigs = {
   [CACHE_TYPES.GLOBAL_PERSISTENT]: {
     persistent: true,
-    ttl: TIME.CACHE_24_HOURS,
+    ttl: TIME.CACHE_1_MINUTE,
     type: 'global_persistent'
   },
 
   [CACHE_TYPES.USER_PERSISTENT]: {
     persistent: true,
-    ttl: TIME.CACHE_12_HOURS, 
+    ttl: TIME.CACHE_1_MINUTE,
     type: 'user_persistent'
   }
 }
