+++ src/services/v2/base/queryClient.js
@@ -4,6 +4,10 @@ import { getCacheConfig } from './cache/CacheConfig'
 import { TIME, RETRY } from './constants'
 
 class EnhancedQueryClient {
+  /**
+   * Inicializa a instância do EnhancedQueryClient
+   * Cria o cacheManager, instancia o QueryClient do TanStack, define flags de inicialização
+   */
   constructor() {
     this.cacheManager = cacheManager
     this.queryClient = this._createQueryClient()
@@ -11,6 +15,11 @@ class EnhancedQueryClient {
     this.initializationPromise = null
   }
 
+  /**
+   * Inicializa o cliente de forma assíncrona com controle de concorrência
+   * Verifica se já foi inicializado, reutiliza promise em andamento ou cria nova inicialização
+   * Evita múltiplas inicializações simultâneas
+   */
   async initialize() {
     if (this.isInitialized) return
 
@@ -22,17 +31,30 @@ class EnhancedQueryClient {
     return this.initializationPromise
   }
 
+  /**
+   * Executa a inicialização real do sistema
+   * Inicializa o cacheManager, configura subscrições de cache e marca como inicializado
+   */
   async _doInitialize() {
     await this.cacheManager.initialize()
     this._setupCacheSubscription()
     this.isInitialized = true
   }
 
+  /**
+   * Inicialização antecipada com retorno de status
+   * Chama initialize() e retorna o status de inicialização
+   */
   async initializeEarly() {
     await this.initialize()
     return this.isInitialized
   }
 
+  /**
+   * Cria e configura o QueryClient do TanStack
+   * Configurações: staleTime (5min), gcTime (10min), retry (2x queries, 1x mutations)
+   * refetchOnWindowFocus: false, refetchOnReconnect: true
+   */
   _createQueryClient() {
     return new QueryClient({
       defaultOptions: {
@@ -50,6 +72,10 @@ class EnhancedQueryClient {
     })
   }
 
+  /**
+   * Configura subscrição para eventos de cache
+   * Monitora queries bem-sucedidas e persiste dados automaticamente
+   */
   _setupCacheSubscription() {
     this.queryClient.getQueryCache().subscribe((event) => {
       if (event.type === 'updated' && event.query.state.status === 'success') {
@@ -58,6 +84,10 @@ class EnhancedQueryClient {
     })
   }
 
+  /**
+   * Processa queries bem-sucedidas para persistência
+   * Verifica se query tem dados e é persistente, serializa chave, obtém configuração e salva no cacheManager
+   */
   async _handleQuerySuccess(event) {
     const { queryKey, state, meta } = event.query
 
@@ -77,6 +107,10 @@ class EnhancedQueryClient {
     }
   }
 
+  /**
+   * Converte chave da query em string serializada
+   * Trata arrays e objetos, ordena chaves para consistência, converte null/undefined para 'null'
+   */
   _serializeQueryKey(queryKey) {
     if (!Array.isArray(queryKey)) {
       return String(queryKey)
@@ -104,6 +138,12 @@ class EnhancedQueryClient {
       .join('_')
   }
 
+  /**
+   * Hook principal para executar queries com suporte a persistência
+   * Com persistência: verifica cache primeiro, executa queryFn se cache expirou
+   * Sem persistência: executa queryFn diretamente
+   * Configurações especiais para queries persistentes: staleTime: 0, refetchOnMount: true, etc.
+   */
   useQuery(options) {
     const { queryKey, queryFn, persistent, ...restOptions } = options
 
@@ -138,6 +178,10 @@ class EnhancedQueryClient {
     })
   }
 
+  /**
+   * Recupera dados do cache se válidos
+   * Inicializa se necessário, serializa chave e busca no cacheManager
+   */
   async _getCachedDataIfValid(queryKey) {
     try {
       await this.initialize()
@@ -149,26 +193,51 @@ class EnhancedQueryClient {
     }
   }
 
+  /**
+   * Invalida queries específicas
+   * Delega para o QueryClient do TanStack
+   */
   invalidateQueries(...args) {
     return this.queryClient.invalidateQueries(...args)
   }
 
+  /**
+   * Define dados diretamente no cache da query
+   * Delega para o QueryClient do TanStack
+   */
   setQueryData(...args) {
     return this.queryClient.setQueryData(...args)
   }
 
+  /**
+   * Recupera dados do cache da query
+   * Delega para o QueryClient do TanStack
+   */
   getQueryData(...args) {
     return this.queryClient.getQueryData(...args)
   }
 
+  /**
+   * Limpa todo o cache persistente
+   * Chama cacheManager.clear()
+   */
   async clearCache() {
     await this.cacheManager.clear()
   }
 
+  /**
+   * Limpa cache persistente de um usuário específico
+   * Limpa cache com escopo de usuário
+   */
   async clearUserPersistentCache(userId) {
     await this.cacheManager.clear({ userScope: userId })
   }
 
+  /**
+   * Pré-carrega dados do cache para uma query
+   * Verifica se dados já existem no QueryClient, se não existem busca no cache persistente
+   * Se encontra no cache, define no QueryClient e retorna dados encontrados
+   */
   async preloadCache(queryKey) {
     await this.initialize()
     const existingData = this.queryClient.getQueryData(queryKey)
